+++ protocols/Weather/weather_ini.cpp
@@ -137,7 +137,7 @@ void WICondListAdd(char *str, WICONDLIST *List)
 {
 	WICONDITEM *newItem = (WICONDITEM*)mir_alloc(sizeof(WICONDITEM));
 	wSetData(&newItem->Item, str);
-	CharLowerBuff(newItem->Item, _tcslen( newItem->Item ));
+	CharLowerBuff(newItem->Item, (DWORD)_tcslen( newItem->Item ));
 	newItem->Next = NULL;
 	if (List->Tail == NULL)	List->Head = newItem;
 	else List->Tail->Next = newItem;
